add a section for default.vue layout
See also https://nuxtjs.org/docs/directory-structure/layouts/#default-layout which is pretty much the same thing I'd like to overhaul the whole doc here actually but this is in my opinion at least worth a mention, especially since the section for app.vue seems to be wrong or at least confusing. But I'll leave that up to you.